change storybook tutorial structure
