+++ packages/core/stories/tutorials/getting-started/angular.ts
@@ -1,0 +1,6 @@
+import marked from "marked";
+
+export const angularTutorial = marked(`
+# Carbon Charts in Angular
+## Getting Started
+`);
