+++ packages/core/stories/tutorials/getting-started/react.ts
@@ -1,0 +1,6 @@
+import marked from "marked";
+
+export const reactTutorial = marked(`
+# Carbon Charts in React
+## Getting Started
+`);
