+++ packages/core/stories/tutorials/getting-started/vanilla.ts
@@ -1,0 +1,6 @@
+import marked from "marked";
+
+export const vanillaTutorial = marked(`
+# Carbon Charts in VanillaJS
+## Getting Started
+`);
