+++ packages/core/stories/tutorials/getting-started/vue.ts
@@ -1,0 +1,6 @@
+import marked from "marked";
+
+export const vueTutorial = marked(`
+# Carbon Charts in Vue
+## Getting Started
+`);
